+++ mtools/mlogfilter/mlogfilter.py
@@ -2,24 +2,27 @@
 
 import argparse, re
 import sys
+import inspect
 
 from mtools.util.logline import LogLine
 from mtools.util.cmdlinetool import LogFileTool
 from mtools.mlogfilter.filters import *
 
-
+import mtools.mlogfilter.filters as filters
 
 class MLogFilterTool(LogFileTool):
 
     def __init__(self):
         LogFileTool.__init__(self, multiple_logfiles=False, stdin_allowed=True)
         
-        self.filters = [] 
+        # add all filter classes from the filters module
+        self.filters = [c[1] for c in inspect.getmembers(filters, inspect.isclass)]
 
         self.argparser.description = 'mongod/mongos log file parser. Use parameters to enable filters. A line only gets printed if it passes all enabled filters.'
         self.argparser.add_argument('--verbose', action='store_true', help='outputs information about the parser and arguments.')
         self.argparser.add_argument('--shorten', action='store', type=int, default=False, nargs='?', metavar='LENGTH', help='shortens long lines by cutting characters out of the middle until the length is <= LENGTH (default 200)')
         self.argparser.add_argument('--exclude', action='store_true', default=False, help='if set, excludes the matching lines rather than includes them.')
+        self.argparser.add_argument('--human', action='store_true', help='outputs numbers formatted with commas and milliseconds as hr,min,sec,ms for easier readability ')
 
 
     def addFilter(self, filterClass):
@@ -34,15 +37,67 @@ class MLogFilterTool(LogFileTool):
         else:
             return arr
 
-    
-    def _outputLine(self, line, length=None):
+    def _outputLine(self, line, length=None, human=False):
         if length:
             if len(line) > length:
-                line = line[:length/2-2] + '...' + line[-length/2:]
+                line = line[:length/2-2] + '...' + line[-length/2+1:]
+        if human:
+            line = self._changeMs(line)
+            line = self._formatNumbers(line)
         print line
 
-    
-    def run(self):
+
+    def _msToString(self, ms):
+        """ changes milliseconds to hours min sec ms format """ 
+        hr, ms = divmod(ms, 3600000)
+        mins, ms = divmod(ms, 60000)
+        secs, mill = divmod(ms, 1000)
+        return "%ihr %imin %isecs %ims"%(hr, mins, secs, mill) 
+
+
+    def _changeMs(self, line):
+        """ changes the ms part in the string if needed """ 
+        # use the the position of the last space instead
+        try:
+            last_space_pos = line.rindex(' ')
+        except ValueError, s:
+            return line
+        else:
+            end_str = line[last_space_pos:]
+            new_string = line
+            if end_str[-2:] == 'ms' and int(end_str[:-2]) >= 1000:
+                # isolate the number of milliseconds 
+                ms = int(end_str[:-2])
+                # create the new string with the beginning part of the log with the new ms part added in
+                new_string = line[:last_space_pos] + ' (' +  self._msToString(ms) + ')' + line[last_space_pos:]
+            return new_string
+
+    def _formatNumbers(self, line):
+        """ formats the numbers so that there are commas inserted, ie. 1200300 becomes 1,200,300 """
+        last_index = 0
+        try:
+            # find the index of the last } character
+            last_index = (line.rindex('}') + 1)
+            end = line[last_index:]
+        except ValueError, e:
+            return line
+        else:
+            # split the string on numbers to isolate them
+            splitted = re.split("(\d+)", end)
+            for index, val in enumerate(splitted):
+                converted = 0
+                try:
+                    converted = int(val)
+                #if it's not an int pass and don't change the string
+                except ValueError, e:
+                    pass
+                else:
+                    if converted > 1000:
+                        splitted[index] = format(converted, ",d")
+            return line[:last_index] + ("").join(splitted)
+
+
+    def run(self, arguments=None):
         """ parses the logfile and asks each filter if it accepts the line.
             it will only be printed if all filters accept the line.
         """
@@ -82,16 +137,15 @@ class MLogFilterTool(LogFileTool):
 
         for line in self.args['logfile']:
             logline = LogLine(line)
-
             if self.args['exclude']:
                 # print line if any filter disagrees
                 if any([not f.accept(logline) for f in self.filters]):
-                    self._outputLine(logline.line_str, self.args['shorten'])
+                    self._outputLine(logline.line_str, self.args['shorten'], self.args['human'])
 
             else:
                 # only print line if all filters agree
                 if all([f.accept(logline) for f in self.filters]):
-                    self._outputLine(logline.line_str, self.args['shorten'])
+                    self._outputLine(logline.line_str, self.args['shorten'], self.args['human'])
 
                 # if at least one filter refuses to accept any remaining lines, stop
                 if any([f.skipRemaining() for f in self.filters]):
@@ -102,24 +156,5 @@ class MLogFilterTool(LogFileTool):
 
 if __name__ == '__main__':
 
-    tool = MLogFilterTool()
-
-    # add filters
-    tool.addFilter(LogLineFilter)
-    tool.addFilter(SlowFilter)
-    tool.addFilter(FastFilter)
-    tool.addFilter(WordFilter)
-    tool.addFilter(TableScanFilter)
-    tool.addFilter(DateTimeFilter)
-    
+    tool = MLogFilterTool()    
     tool.run()
-
-
-
-
-
-
-
-
-
-