+++ mtools/mplotqueries/plottypes/__init__.py
@@ -1,3 +1,5 @@
 from scatter_type import ScatterPlotType, DurationPlotType, NScannedNPlotType
 from event_type import EventPlotType, RSStatePlotType
 from range_type import RangePlotType
+from histogram_type import HistogramPlotType
+from connchurn_type import ConnectionChurnPlotType